homework oprators

+++ 12 feb 24/homework/Shortprogram.java
@@ -3,7 +3,7 @@ class Shortprogram{
     public static void main(String[] args){
 
         short a = 6;
-        short b = 4;
+        long b = 4;
 
         System.out.println("The value in int is  "+   (int)(a +b));
         System.out.println("The value in byte is "+  (byte)(a-b));
